Cleaned out extra jolicloud-desktop packages not shipped with the ISO
Since printing and other features were removed from the PreBeta, they were
taken out of the ISO but not the jolicloud-desktop metapackage. As a
result, jolicloud-desktop itself was accidentally removed.

This made it impossible to ship updated packages via jolicloud-desktop to
PreBeta installed netbooks.

The solution is to make one of hte existing packages (jolicloud-minimal)
depend on jolicloud-standard, but this installed the packages that were
previously removed. This commit removes those unnecessary packages from the
master meta-package.